fix a-frame geometry defaults

+++ Runtime/ArenaMesh/ArenaMeshTorusKnot.cs
@@ -11,8 +11,8 @@ namespace ArenaUnity
         [SerializeField, Range(0.05f, 10f)] internal float radiusTubular = 0.1f;
         [SerializeField, Range(2, 64)] internal int radialSegments = 16;
         [SerializeField, Range(3, 64)] internal int thetaSegments = 8;
-        [SerializeField, Range(2, 5)] internal int p = 2;
-        [SerializeField, Range(2, 5)] internal int q = 3;
+        [SerializeField, Range(2, 5)] internal float p = 2;
+        [SerializeField, Range(2, 5)] internal float q = 3;
 
         protected override void Build(MeshFilter filter)
         {
